refactor(tui::ui::components::popups::help): move function "mount_help_popup" to popup file
